Move utils.pcap to lib.cuckoo.common.pcap_utils
